% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pseudo_order.R
\name{build_pseudo_order}
\alias{build_pseudo_order}
\title{build_pseudo_order
This function is used to build a quasi-timing sequence using
a number of different methods.}
\usage{
build_pseudo_order(
  data,
  reverse = FALSE,
  method = c("Default", "tricycle", "reCAT")
)
}
\arguments{
\item{data}{Expression profile data of cell cycle genes}

\item{reverse}{Whether to change the direction of pseudo order}

\item{method}{Which method to use}
}
\value{

}
\description{
build_pseudo_order
This function is used to build a quasi-timing sequence using
a number of different methods.
}
