% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pseudo_order.R
\name{cor_scatter_plot}
\alias{cor_scatter_plot}
\title{cor_scatter_plot}
\usage{
cor_scatter_plot(index_data)
}
\arguments{
\item{index_data}{a data frame.}
}
\value{

}
\description{
This function is used to plot a scatter plot that maximizes correlation
}
