% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{pca_plot}
\alias{pca_plot}
\title{pca_plot}
\usage{
pca_plot(data, time.iqr = 1)
}
\arguments{
\item{data}{Expression matrix of cell cycle genes}

\item{time.iqr}{times of IQR, defualt is 1.5}
}
\value{

}
\description{
This function is used to draw the distribution map of each point
on pc1 and pc2 after the first step of pca, and is used to screen
and remove outliers
}
