% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{remove_pc_outliers}
\alias{remove_pc_outliers}
\title{remove_pc_outliers}
\usage{
remove_pc_outliers(pc_score, time.iqr = 1)
}
\arguments{
\item{pc_score}{A dataframe containing scores on PC1 and PC2}
}
\value{

}
\description{
remove_pc_outliers
}
