% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pseudo_order.R
\name{rings2aligment}
\alias{rings2aligment}
\title{rings2aligment}
\usage{
rings2aligment(input_vector, reference_vector)
}
\arguments{
\item{input_vector}{a vector}

\item{reference_vector}{another vector}
}
\value{
output_posi, output_coef
}
\description{
This function is used to align the ring structure
Adjust the order position of the elements in input_vector,
Maximizes the Spearman correlation coefficients of input_vector and reference_vector
}
\examples{
v1=c(0.6,0.7,0.8,0.1,0.2,0.3,0.4,0.5)
v2=c(1:8)
res <- rings2aligment(v1, v2)
res
$output_posi
[1] 4 5 6 7 8 1 2 3

$output_coef
[1] 1
v1[res$output_posi]
}
